requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
duckdb>=0.9
psycopg[binary]>=3.1
PyJWT>=2.9
